+++ Build-Parallel.ps1
@@ -12,7 +12,7 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 #>
 
-#.ExternalHelp Invoke-Build-Help.xml
+#.ExternalHelp InvokeBuild-Help.xml
 param(
 	[Parameter(Position=0)][hashtable[]]$Build,
 	$Result,